Add (Try)GetDateTime(Offset) to JsonElement
