+++ src/System.Text.Json/src/System/Text/Json/Reader/JsonReaderHelper.cs
@@ -48,10 +48,6 @@ namespace System.Text.Json
             }
         }
 
-        // A digit is valid if it is in the range: [0..9]
-        // Otherwise, return false.
-        public static bool IsDigit(byte nextByte) => (uint)(nextByte - '0') <= '9' - '0';
-
         // Returns true if the TokenType is a primitive "value", i.e. String, Number, True, False, and Null
         // Otherwise, return false.
         public static bool IsTokenTypePrimitive(JsonTokenType tokenType) =>
